Consolidated GridProtocol data into Grid.swift.
