+++ Assignment4/Assignment4/Grid.swift
@@ -5,6 +5,49 @@ import Foundation
 
 fileprivate func norm(_ val: Int, to size: Int) -> Int { return ((val % size) + size) % size }
 
+public struct GridPosition: Equatable {
+    var row: Int
+    var col: Int
+    
+    public static func == (lhs: GridPosition, rhs: GridPosition) -> Bool {
+        return (lhs.row == rhs.row && lhs.col == rhs.col)
+    }
+}
+
+public struct GridSize {
+    var rows: Int
+    var cols: Int
+}
+
+public enum CellState {
+    case alive, empty, born, died
+    
+    public var isAlive: Bool {
+        switch self {
+        case .alive, .born: return true
+        default: return false
+        }
+    }
+    func description() -> String {
+        switch self {
+        case .alive: return "alive"
+        case .empty: return "empty"
+        case .born: return "born"
+        case .died: return "died"
+        }
+    }
+}
+
+public protocol GridViewDataSource {
+    subscript (row: Int, col: Int) -> CellState { get set }
+}
+
+public protocol GridProtocol: CustomStringConvertible {
+    init(_ size: GridSize, cellInitializer: (GridPosition) -> CellState)
+    var size: GridSize { get }
+    func next() -> Self
+}
+
 public let lazyPositions = { (size: GridSize) in
     return (0 ..< size.rows)
         .lazy
@@ -130,6 +173,15 @@ public extension Grid {
 
 protocol EngineDelegate {
     func engineDidUpdate(engine: StandardEngine)
+}
+
+protocol EngineProtocol {
+    var grid: GridProtocol { get }
+    var refreshRate: Double { get set }
+    var refreshTimer: Timer { get set }
+    var rows: Int { get set }
+    var cols: Int { get set }
+    
 }
 
 class StandardEngine {
